requirements.txt
# Audio processing
audioop-lts==0.2.2
vosk>=0.3.45
sounddevice==0.4.6
soundfile==0.12.1
pydub==0.25.1
pyperclip==1.8.2
pyttsx3==2.90

# GUI
customtkinter==5.2.0
Pillow>=10.3.0

# Text processing and NLP
nltk==3.8.1

# Document generation
python-docx==1.1.0
reportlab==4.0.7

# Numerical computing
numpy>=1.26.0

# Dependencies for model download
requests>=2.31.0
tqdm>=4.66.0

# Standard libraries compatibility
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0

# Build and installation tools
setuptools
